+++ component/backend/View/ControlPanel/Html.php
@@ -16,7 +16,6 @@ defined('_JEXEC') or die;
 
 class Html extends \FOF30\View\DataView\Html
 {
-	public $hasGeoIPPlugin;
 	public $akeebaCommonDatePHP;
 	public $akeebaCommonDateObsolescence;
 
@@ -24,8 +23,6 @@ class Html extends \FOF30\View\DataView\Html
 	{
 		/** @var ControlPanel $model */
 		$model = $this->getModel();
-
-		$this->hasGeoIPPlugin = $model->hasGeoIPPlugin();
 
 		$this->akeebaCommonDatePHP = $this->container->platform->getDate('2015-08-14 00:00:00', 'GMT')->format(JText::_('DATE_FORMAT_LC1'));
 		$this->akeebaCommonDateObsolescence = $this->container->platform->getDate('2016-05-14 00:00:00', 'GMT')->format(JText::_('DATE_FORMAT_LC1'));